@@ -16,7 +16,13 @@
         "forum": "https://discourse.laminas.dev"
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "platform": {
+            "php": "7.4.99"
+        },
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
     },
     "extra": {
         "laminas": {
@@ -25,7 +31,7 @@
         }
     },
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0",
+        "php": "^7.4 || ~8.0.0 || ~8.1.0",
         "ext-intl": "*",
         "laminas/laminas-stdlib": "^2.7 || ^3.0"
     },
@@ -33,13 +39,14 @@
         "laminas/laminas-cache": "^3.1.2",
         "laminas/laminas-cache-storage-adapter-memory": "^2.0.0",
         "laminas/laminas-cache-storage-deprecated-factory": "^1.0.0",
-        "laminas/laminas-coding-standard": "~1.0.0",
+        "laminas/laminas-coding-standard": "~2.3.0",
         "laminas/laminas-config": "^3.4.0",
         "laminas/laminas-eventmanager": "^3.4.0",
         "laminas/laminas-filter": "^2.10.0",
         "laminas/laminas-servicemanager": "^3.7.0",
         "laminas/laminas-validator": "^2.14.0",
         "laminas/laminas-view": "^2.12.0",
+        "phpspec/prophecy-phpunit": "^2.0",
         "phpunit/phpunit": "^9.3",
         "psalm/plugin-phpunit": "^0.16.1",
         "vimeo/psalm": "^4.21"
