@@ -14,16 +14,15 @@ class TranslatorServiceFactory implements FactoryInterface
     /**
      * Create a Translator instance.
      *
-     * @param ContainerInterface $container
      * @param string $requestedName
      * @param null|array $options
      * @return Translator
      */
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null)
     {
         // Configure the translator
         $config     = $container->get('config');
-        $trConfig   = isset($config['translator']) ? $config['translator'] : [];
+        $trConfig   = $config['translator'] ?? [];
         $translator = Translator::factory($trConfig);
         if ($container->has('TranslatorPluginManager')) {
             $translator->setPluginManager($container->get('TranslatorPluginManager'));
@@ -36,7 +35,6 @@ class TranslatorServiceFactory implements FactoryInterface
      *
      * Proxies to `__invoke()`.
      *
-     * @param ServiceLocatorInterface $serviceLocator
      * @return Translator
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
