@@ -1,69 +1,48 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\Filter;
 
 use Laminas\I18n\Filter\Alnum as AlnumFilter;
+use LaminasTest\I18n\TestCase;
 use Locale;
-use PHPUnit\Framework\TestCase;
+use stdClass;
 
-/**
- * @group      Laminas_Filter
- */
+use function array_keys;
+use function array_values;
+use function preg_match;
+
 class AlnumTest extends TestCase
 {
-    /**
-     * AlnumFilter object
-     *
-     * @var AlnumFilter
-     */
-    protected $filter;
+    private AlnumFilter $filter;
 
     /**
      * Is PCRE is compiled with UTF-8 and Unicode support
-     *
-     * @var mixed
-     **/
-    protected static $unicodeEnabled;
-
-    /**
-     * Locale in browser.
-     *
-     * @var string object
      */
-    protected $locale;
+    protected static bool $unicodeEnabled;
 
     /**
      * The Alphabet means english alphabet.
-     *
-     * @var bool
      */
-    protected static $meansEnglishAlphabet;
+    protected static bool $meansEnglishAlphabet;
 
     /**
      * Creates a new AlnumFilter object for each test method
-     *
-     * @return void
      */
     protected function setUp(): void
     {
-        if (! extension_loaded('intl')) {
-            $this->markTestSkipped('ext/intl not enabled');
-        }
-
-        $this->filter = new AlnumFilter();
-
-        $this->locale               = Locale::getDefault();
-        $language                   = Locale::getPrimaryLanguage($this->locale);
+        parent::setUp();
+        $this->filter                 = new AlnumFilter();
+        $language                     = Locale::getPrimaryLanguage(Locale::getDefault());
         static::$meansEnglishAlphabet = $language === 'ja';
         static::$unicodeEnabled       = (bool) @preg_match('/\pL/u', 'a');
     }
 
     /**
      * Ensures that the filter follows expected behavior
-     *
-     * @return void
      */
-    public function testBasic()
+    public function testBasic(): void
     {
         if (! static::$unicodeEnabled) {
             // POSIX named classes are not supported, use alternative a-zA-Z match
@@ -72,7 +51,7 @@ class AlnumTest extends TestCase
                 'abc 123' => 'abc123',
                 'abcxyz'  => 'abcxyz',
                 'AZ@#4.3' => 'AZ43',
-                ''        => ''
+                ''        => '',
             ];
         } elseif (static::$meansEnglishAlphabet) {
             // The Alphabet means english alphabet.
@@ -85,20 +64,20 @@ class AlnumTest extends TestCase
              * The third  contains various multibyte or singebyte characters.
              */
             $valuesExpected = [
-                'aＡBｂ3４5６'  => 'aB35',
-                'z７ Ｙ8　x９'  => 'z8x',
+                'aＡBｂ3４5６'    => 'aB35',
+                'z７ Ｙ8　x９'    => 'z8x',
                 '，s1.2r３#:q,' => 's12rq',
             ];
         } else {
             //The Alphabet means each language's alphabet.
             $valuesExpected = [
-                'abc123'        => 'abc123',
-                'abc 123'       => 'abc123',
-                'abcxyz'        => 'abcxyz',
-                'če2t3ně'       => 'če2t3ně',
-                'grz5e4gżółka'  => 'grz5e4gżółka',
-                'Be3l5gië'      => 'Be3l5gië',
-                ''              => ''
+                'abc123'       => 'abc123',
+                'abc 123'      => 'abc123',
+                'abcxyz'       => 'abcxyz',
+                'če2t3ně'      => 'če2t3ně',
+                'grz5e4gżółka' => 'grz5e4gżółka',
+                'Be3l5gië'     => 'Be3l5gië',
+                ''             => '',
             ];
         }
 
@@ -110,10 +89,8 @@ class AlnumTest extends TestCase
 
     /**
      * Ensures that the allowWhiteSpace option works as expected
-     *
-     * @return void
      */
-    public function testAllowWhiteSpace()
+    public function testAllowWhiteSpace(): void
     {
         $this->filter->setAllowWhiteSpace(true);
 
@@ -126,13 +103,13 @@ class AlnumTest extends TestCase
                 'AZ@#4.3' => 'AZ43',
                 ''        => '',
                 "\n"      => "\n",
-                " \t "    => " \t "
+                " \t "    => " \t ",
             ];
         } elseif (static::$meansEnglishAlphabet) {
             //The Alphabet means english alphabet.
             $valuesExpected = [
                 'a B ４5' => 'a B 5',
-                'z3　x'   => 'z3x'
+                'z3　x'   => 'z3x',
             ];
         } else {
             //The Alphabet means each language's alphabet.
@@ -153,7 +130,7 @@ class AlnumTest extends TestCase
         }
     }
 
-    public function testFilterSupportArray()
+    public function testFilterSupportArray(): void
     {
         $filter = new AlnumFilter();
 
@@ -162,7 +139,7 @@ class AlnumTest extends TestCase
             'abc 123' => 'abc123',
             'abcxyz'  => 'abcxyz',
             'AZ@#4.3' => 'AZ43',
-            ''        => ''
+            ''        => '',
         ];
 
         $actual = $filter->filter(array_keys($values));
@@ -170,19 +147,20 @@ class AlnumTest extends TestCase
         $this->assertEquals(array_values($values), $actual);
     }
 
-    public function returnUnfilteredDataProvider()
+    /** @return array<array-key, array{0: mixed}> */
+    public function returnUnfilteredDataProvider(): array
     {
         return [
             [null],
-            [new \stdClass()]
+            [new stdClass()],
         ];
     }
 
     /**
      * @dataProvider returnUnfilteredDataProvider
-     * @return void
+     * @param mixed $input
      */
-    public function testReturnUnfiltered($input)
+    public function testReturnUnfiltered($input): void
     {
         $filter = new AlnumFilter();
 
