@@ -1,20 +1,29 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\Translator\Loader;
 
+use Laminas\I18n\Exception\InvalidArgumentException;
 use Laminas\I18n\Translator\Loader\Ini as IniLoader;
-use PHPUnit\Framework\TestCase;
+use Laminas\I18n\Translator\TextDomain;
+use LaminasTest\I18n\TestCase;
+
+use function get_include_path;
+use function realpath;
+use function set_include_path;
+
+use const PATH_SEPARATOR;
 
 class IniTest extends TestCase
 {
-    protected $testFilesDir;
-    protected $originalLocale;
-    protected $originalIncludePath;
+    private string $testFilesDir;
+    private string $originalIncludePath;
 
     protected function setUp(): void
     {
-        $this->testFilesDir = realpath(__DIR__ . '/../_files');
-
+        parent::setUp();
+        $this->testFilesDir        = realpath(__DIR__ . '/../_files');
         $this->originalIncludePath = get_include_path();
         set_include_path($this->testFilesDir . PATH_SEPARATOR . $this->testFilesDir . '/translations.phar');
     }
@@ -22,67 +31,68 @@ class IniTest extends TestCase
     protected function tearDown(): void
     {
         set_include_path($this->originalIncludePath);
+        parent::tearDown();
     }
 
-    public function testLoaderFailsToLoadMissingFile()
+    public function testLoaderFailsToLoadMissingFile(): void
     {
         $loader = new IniLoader();
-        $this->expectException('Laminas\I18n\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Could not find or open file');
         $loader->load('en_EN', 'missing');
     }
 
-    public function testLoaderLoadsEmptyFile()
+    public function testLoaderLoadsEmptyFile(): void
     {
-        $loader = new IniLoader();
+        $loader     = new IniLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_empty.ini');
-        $this->assertInstanceOf('Laminas\I18n\Translator\TextDomain', $textDomain);
+        $this->assertInstanceOf(TextDomain::class, $textDomain);
     }
 
-    public function testLoaderFailsToLoadNonArray()
+    public function testLoaderFailsToLoadNonArray(): void
     {
         $loader = new IniLoader();
-        $this->expectException('Laminas\I18n\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Each INI row must be an array with message and translation');
         $loader->load('en_EN', $this->testFilesDir . '/failed.ini');
     }
 
-    public function testLoaderFailsToLoadBadSyntax()
+    public function testLoaderFailsToLoadBadSyntax(): void
     {
         $loader = new IniLoader();
-        $this->expectException('Laminas\I18n\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Each INI row must be an array with message and translation');
         $loader->load('en_EN', $this->testFilesDir . '/failed_syntax.ini');
     }
 
-    public function testLoaderReturnsValidTextDomain()
+    public function testLoaderReturnsValidTextDomain(): void
     {
-        $loader = new IniLoader();
+        $loader     = new IniLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_en.ini');
 
         $this->assertEquals('Message 1 (en)', $textDomain['Message 1']);
         $this->assertEquals('Message 4 (en)', $textDomain['Message 4']);
     }
 
-    public function testLoaderReturnsValidTextDomainWithFileWithoutPlural()
+    public function testLoaderReturnsValidTextDomainWithFileWithoutPlural(): void
     {
-        $loader = new IniLoader();
+        $loader     = new IniLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_en_without_plural.ini');
 
         $this->assertEquals('Message 1 (en)', $textDomain['Message 1']);
         $this->assertEquals('Message 4 (en)', $textDomain['Message 4']);
     }
 
-    public function testLoaderReturnsValidTextDomainWithSimpleSyntax()
+    public function testLoaderReturnsValidTextDomainWithSimpleSyntax(): void
     {
-        $loader = new IniLoader();
+        $loader     = new IniLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_en_simple_syntax.ini');
 
         $this->assertEquals('Message 1 (en)', $textDomain['Message 1']);
         $this->assertEquals('Message 4 (en)', $textDomain['Message 4']);
     }
 
-    public function testLoaderLoadsPluralRules()
+    public function testLoaderLoadsPluralRules(): void
     {
         $loader     = new IniLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_en.ini');
@@ -93,7 +103,7 @@ class IniTest extends TestCase
         $this->assertEquals(2, $textDomain->getPluralRule()->evaluate(10));
     }
 
-    public function testLoaderLoadsFromIncludePath()
+    public function testLoaderLoadsFromIncludePath(): void
     {
         $loader = new IniLoader();
         $loader->setUseIncludePath(true);
@@ -103,7 +113,7 @@ class IniTest extends TestCase
         $this->assertEquals('Message 4 (en)', $textDomain['Message 4']);
     }
 
-    public function testLoaderLoadsFromPhar()
+    public function testLoaderLoadsFromPhar(): void
     {
         $loader = new IniLoader();
         $loader->setUseIncludePath(true);
