@@ -1,73 +1,75 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\Translator\Loader;
 
+use Laminas\I18n\Exception\InvalidArgumentException;
 use Laminas\I18n\Translator\Loader\PhpArray as PhpArrayLoader;
+use Laminas\I18n\Translator\TextDomain;
+use LaminasTest\I18n\TestCase;
 use Locale;
-use PHPUnit\Framework\TestCase;
+
+use function get_include_path;
+use function realpath;
+use function set_include_path;
+
+use const PATH_SEPARATOR;
 
 class PhpArrayTest extends TestCase
 {
-    protected $testFilesDir;
-    protected $originalLocale;
-    protected $originalIncludePath;
+    private string $testFilesDir;
+    private string $originalIncludePath;
 
     protected function setUp(): void
     {
-        if (! extension_loaded('intl')) {
-            $this->markTestSkipped('ext/intl not enabled');
-        }
-
-        $this->originalLocale = Locale::getDefault();
+        parent::setUp();
         Locale::setDefault('en_EN');
 
-        $this->testFilesDir = realpath(__DIR__ . '/../_files');
-
+        $this->testFilesDir        = realpath(__DIR__ . '/../_files');
         $this->originalIncludePath = get_include_path();
         set_include_path($this->testFilesDir . PATH_SEPARATOR . $this->testFilesDir . '/translations.phar');
     }
 
     protected function tearDown(): void
     {
-        if (extension_loaded('intl')) {
-            Locale::setDefault($this->originalLocale);
-            set_include_path($this->originalIncludePath);
-        }
+        set_include_path($this->originalIncludePath);
+        parent::tearDown();
     }
 
-    public function testLoaderFailsToLoadMissingFile()
+    public function testLoaderFailsToLoadMissingFile(): void
     {
         $loader = new PhpArrayLoader();
-        $this->expectException('Laminas\I18n\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Could not find or open file');
         $loader->load('en_EN', 'missing');
     }
 
-    public function testLoaderFailsToLoadNonArray()
+    public function testLoaderFailsToLoadNonArray(): void
     {
         $loader = new PhpArrayLoader();
-        $this->expectException('Laminas\I18n\Exception\InvalidArgumentException');
+        $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessage('Expected an array, but received');
         $loader->load('en_EN', $this->testFilesDir . '/failed.php');
     }
 
-    public function testLoaderLoadsEmptyArray()
+    public function testLoaderLoadsEmptyArray(): void
     {
-        $loader = new PhpArrayLoader();
+        $loader     = new PhpArrayLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_empty.php');
-        $this->assertInstanceOf('Laminas\I18n\Translator\TextDomain', $textDomain);
+        $this->assertInstanceOf(TextDomain::class, $textDomain);
     }
 
-    public function testLoaderReturnsValidTextDomain()
+    public function testLoaderReturnsValidTextDomain(): void
     {
-        $loader = new PhpArrayLoader();
+        $loader     = new PhpArrayLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_en.php');
 
         $this->assertEquals('Message 1 (en)', $textDomain['Message 1']);
         $this->assertEquals('Message 4 (en)', $textDomain['Message 4']);
     }
 
-    public function testLoaderLoadsPluralRules()
+    public function testLoaderLoadsPluralRules(): void
     {
         $loader     = new PhpArrayLoader();
         $textDomain = $loader->load('en_EN', $this->testFilesDir . '/translation_en.php');
@@ -78,7 +80,7 @@ class PhpArrayTest extends TestCase
         $this->assertEquals(2, $textDomain->getPluralRule()->evaluate(10));
     }
 
-    public function testLoaderLoadsFromIncludePath()
+    public function testLoaderLoadsFromIncludePath(): void
     {
         $loader = new PhpArrayLoader();
         $loader->setUseIncludePath(true);
@@ -88,7 +90,7 @@ class PhpArrayTest extends TestCase
         $this->assertEquals('Message 4 (en)', $textDomain['Message 4']);
     }
 
-    public function testLoaderLoadsFromPhar()
+    public function testLoaderLoadsFromPhar(): void
     {
         $loader = new PhpArrayLoader();
         $loader->setUseIncludePath(true);
