@@ -1,171 +1,168 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\Translator;
 
+use Laminas\Cache\Storage\StorageInterface;
 use Laminas\Cache\StorageFactory as CacheFactory;
+use Laminas\EventManager\Event;
 use Laminas\EventManager\EventInterface;
 use Laminas\I18n\Translator\TextDomain;
 use Laminas\I18n\Translator\Translator;
 use Laminas\ServiceManager\Config;
+use LaminasTest\I18n\TestCase;
 use LaminasTest\I18n\Translator\TestAsset\Loader as TestLoader;
 use Locale;
-use PHPUnit\Framework\TestCase;
 
 class TranslatorTest extends TestCase
 {
-    /**
-     * @var Translator
-     */
-    protected $translator;
-
-    /**
-     * @var string
-     */
-    protected $originalLocale;
-
-    /**
-     * @var string
-     */
-    protected $testFilesDir;
+    private Translator $translator;
+    private string $testFilesDir;
 
     protected function setUp(): void
     {
-        if (! extension_loaded('intl')) {
-            $this->markTestSkipped('ext/intl not enabled');
-        }
-
-        $this->originalLocale = Locale::getDefault();
-        $this->translator     = new Translator();
-
+        parent::setUp();
+        $this->translator = new Translator();
         Locale::setDefault('en_EN');
-
         $this->testFilesDir = __DIR__ . '/_files';
     }
 
-    protected function tearDown(): void
-    {
-        if (extension_loaded('intl')) {
-            Locale::setDefault($this->originalLocale);
-        }
-    }
-
-    public function testFactoryCreatesTranslator()
+    public function testFactoryCreatesTranslator(): void
     {
         $translator = Translator::factory([
-            'locale' => 'de_DE',
+            'locale'   => 'de_DE',
             'patterns' => [
                 [
-                    'type' => 'phparray',
+                    'type'     => 'phparray',
                     'base_dir' => $this->testFilesDir . '/testarray',
-                    'pattern' => 'translation-%s.php'
-                ]
+                    'pattern'  => 'translation-%s.php',
+                ],
             ],
-            'files' => [
+            'files'    => [
                 [
-                    'type' => 'phparray',
+                    'type'     => 'phparray',
                     'filename' => $this->testFilesDir . '/translation_en.php',
-                ]
-            ]
+                ],
+            ],
         ]);
 
-        $this->assertInstanceOf('Laminas\I18n\Translator\Translator', $translator);
+        $this->assertInstanceOf(Translator::class, $translator);
         $this->assertEquals('de_DE', $translator->getLocale());
     }
 
-    public function testTranslationFromSeveralTranslationFiles()
+    public function testTranslationFromSeveralTranslationFiles(): void
     {
         $translator = Translator::factory([
-            'locale' => 'de_DE',
+            'locale'                    => 'de_DE',
             'translation_file_patterns' => [
                 [
-                    'type' => 'phparray',
+                    'type'     => 'phparray',
                     'base_dir' => $this->testFilesDir . '/testarray',
-                    'pattern' => 'translation-%s.php'
+                    'pattern'  => 'translation-%s.php',
                 ],
                 [
-                    'type' => 'phparray',
+                    'type'     => 'phparray',
                     'base_dir' => $this->testFilesDir . '/testarray',
-                    'pattern' => 'translation-more-%s.php'
-                ]
-            ]
+                    'pattern'  => 'translation-more-%s.php',
+                ],
+            ],
         ]);
 
         //Test translator instance
-        $this->assertInstanceOf('Laminas\I18n\Translator\Translator', $translator);
+        $this->assertInstanceOf(Translator::class, $translator);
 
         //Test translations
-        $this->assertEquals('Nachricht 1', $translator->translate('Message 1')); //translation-de_DE.php
-        $this->assertEquals('Nachricht 9', $translator->translate('Message 9')); //translation-more-de_DE.php
-        $this->assertEquals('Nachricht 10 - 0', $translator->translatePlural('Message 10', 'Message 10', 1)); //translation-de_DE.php
-        $this->assertEquals('Nachricht 10 - 1', $translator->translatePlural('Message 10', 'Message 10', 2)); //translation-de_DE.php
-        $this->assertEquals('Nachricht 11 - 0', $translator->translatePlural('Message 11', 'Message 11', 1)); //translation-more-de_DE.php
-        $this->assertEquals('Nachricht 11 - 1', $translator->translatePlural('Message 11', 'Message 11', 2)); //translation-more-de_DE.php
+        $this->assertEquals(
+            'Nachricht 1',
+            $translator->translate('Message 1')
+        ); //translation-de_DE.php
+        $this->assertEquals(
+            'Nachricht 9',
+            $translator->translate('Message 9')
+        ); //translation-more-de_DE.php
+        $this->assertEquals(
+            'Nachricht 10 - 0',
+            $translator->translatePlural('Message 10', 'Message 10', 1)
+        ); //translation-de_DE.php
+        $this->assertEquals(
+            'Nachricht 10 - 1',
+            $translator->translatePlural('Message 10', 'Message 10', 2)
+        ); //translation-de_DE.php
+        $this->assertEquals(
+            'Nachricht 11 - 0',
+            $translator->translatePlural('Message 11', 'Message 11', 1)
+        ); //translation-more-de_DE.php
+        $this->assertEquals(
+            'Nachricht 11 - 1',
+            $translator->translatePlural('Message 11', 'Message 11', 2)
+        ); //translation-more-de_DE.php
     }
 
-    public function testTranslationFromDifferentSourceTypes()
+    public function testTranslationFromDifferentSourceTypes(): void
     {
         $translator = Translator::factory([
-            'locale' => 'de_DE',
+            'locale'                    => 'de_DE',
             'translation_file_patterns' => [
                 [
                     'type'     => 'phparray',
                     'base_dir' => $this->testFilesDir . '/testarray',
-                    'pattern'  => 'translation-de_DE.php'
+                    'pattern'  => 'translation-de_DE.php',
                 ],
             ],
-            'translation_files' => [
+            'translation_files'         => [
                 [
                     'type'     => 'phparray',
-                    'filename' => $this->testFilesDir . '/testarray/translation-more-de_DE.php'
-                ]
-            ]
+                    'filename' => $this->testFilesDir . '/testarray/translation-more-de_DE.php',
+                ],
+            ],
         ]);
 
         $this->assertEquals('Nachricht 1', $translator->translate('Message 1')); //translation-de_DE.php
         $this->assertEquals('Nachricht 9', $translator->translate('Message 9')); //translation-more-de_DE.php
     }
 
-    public function testFactoryCreatesTranslatorWithCache()
+    public function testFactoryCreatesTranslatorWithCache(): void
     {
         $translator = Translator::factory([
-            'locale' => 'de_DE',
+            'locale'   => 'de_DE',
             'patterns' => [
                 [
-                    'type' => 'phparray',
+                    'type'     => 'phparray',
                     'base_dir' => $this->testFilesDir . '/testarray',
-                    'pattern' => 'translation-%s.php'
-                ]
+                    'pattern'  => 'translation-%s.php',
+                ],
             ],
-            'cache' => [
-                'adapter' => 'memory'
-            ]
+            'cache'    => [
+                'adapter' => 'memory',
+            ],
         ]);
 
-        $this->assertInstanceOf('Laminas\I18n\Translator\Translator', $translator);
-        $this->assertInstanceOf('Laminas\Cache\Storage\StorageInterface', $translator->getCache());
+        $this->assertInstanceOf(Translator::class, $translator);
+        $this->assertInstanceOf(StorageInterface::class, $translator->getCache());
     }
 
-    public function testDefaultLocale()
+    public function testDefaultLocale(): void
     {
         $this->assertEquals('en_EN', $this->translator->getLocale());
     }
 
-    public function testForcedLocale()
+    public function testForcedLocale(): void
     {
         $this->translator->setLocale('de_DE');
         $this->assertEquals('de_DE', $this->translator->getLocale());
     }
 
-    public function testTranslate()
+    public function testTranslate(): void
     {
-        $loader = new TestLoader();
+        $loader             = new TestLoader();
         $loader->textDomain = new TextDomain(['foo' => 'bar']);
-        $config = new Config([
+        $config             = new Config([
             'services' => [
-                'test' => $loader
-            ]
+                'test' => $loader,
+            ],
         ]);
-        $pm = $this->translator->getPluginManager();
+        $pm                 = $this->translator->getPluginManager();
         $config->configureServiceManager($pm);
         $this->translator->setPluginManager($pm);
         $this->translator->addTranslationFile('test', null);
@@ -173,7 +170,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('bar', $this->translator->translate('foo'));
     }
 
-    public function testTranslationsLoadedFromCache()
+    public function testTranslationsLoadedFromCache(): void
     {
         $cache = CacheFactory::factory(['adapter' => 'memory']);
         $this->translator->setCache($cache);
@@ -186,15 +183,15 @@ class TranslatorTest extends TestCase
         $this->assertEquals('bar', $this->translator->translate('foo'));
     }
 
-    public function testTranslationsAreStoredInCache()
+    public function testTranslationsAreStoredInCache(): void
     {
         $cache = CacheFactory::factory(['adapter' => 'memory']);
         $this->translator->setCache($cache);
 
-        $loader = new TestLoader();
+        $loader             = new TestLoader();
         $loader->textDomain = new TextDomain(['foo' => 'bar']);
-        $config = new Config(['services' => ['test' => $loader]]);
-        $plugins = $this->translator->getPluginManager();
+        $config             = new Config(['services' => ['test' => $loader]]);
+        $plugins            = $this->translator->getPluginManager();
         $config->configureServiceManager($plugins);
         $this->translator->setPluginManager($plugins);
         $this->translator->addTranslationFile('test', null);
@@ -202,11 +199,11 @@ class TranslatorTest extends TestCase
         $this->assertEquals('bar', $this->translator->translate('foo'));
 
         $item = $cache->getItem($this->translator->getCacheId('default', 'en_EN'));
-        $this->assertInstanceOf('Laminas\I18n\Translator\TextDomain', $item);
+        $this->assertInstanceOf(TextDomain::class, $item);
         $this->assertEquals('bar', $item['foo']);
     }
 
-    public function testTranslationsAreClearedFromCache()
+    public function testTranslationsAreClearedFromCache(): void
     {
         $textDomain = 'default';
         $locale     = 'en_EN';
@@ -226,12 +223,12 @@ class TranslatorTest extends TestCase
         $this->assertFalse($success);
     }
 
-    public function testClearCacheReturnsFalseIfNoCacheIsPresent()
+    public function testClearCacheReturnsFalseIfNoCacheIsPresent(): void
     {
         $this->assertFalse($this->translator->clearCache('default', 'en_EN'));
     }
 
-    public function testTranslatePlurals()
+    public function testTranslatePlurals(): void
     {
         $this->translator->setLocale('en_EN');
         $this->translator->addTranslationFile(
@@ -250,7 +247,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 5 (en) Plural 2', $pl2);
     }
 
-    public function testTranslatePluralsNonExistantLocale()
+    public function testTranslatePluralsNonExistentLocale(): void
     {
         $this->translator->addTranslationFilePattern(
             'phparray',
@@ -269,7 +266,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 5 Plural', $pl2);
     }
 
-    public function testTranslatePluralsNonExistantTranslation()
+    public function testTranslatePluralsNonExistentTranslation(): void
     {
         $this->translator->addTranslationFilePattern(
             'phparray',
@@ -288,7 +285,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 12 Plural', $pl2);
     }
 
-    public function testTranslateNoPlurals()
+    public function testTranslateNoPlurals(): void
     {
         // Some languages such as Japanese and Chinese does not have plural forms
         $this->translator->setLocale('ja_JP');
@@ -308,7 +305,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 9 (ja)', $pl2);
     }
 
-    public function testTranslateNonExistantLocale()
+    public function testTranslateNonExistentLocale(): void
     {
         $this->translator->addTranslationFilePattern(
             'phparray',
@@ -329,7 +326,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 9', $this->translator->translate('Message 9'));
     }
 
-    public function testTranslateNonExistantTranslation()
+    public function testTranslateNonExistentTranslation(): void
     {
         $this->translator->addTranslationFilePattern(
             'phparray',
@@ -344,7 +341,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 13', $this->translator->translate('Message 13'));
     }
 
-    public function testEnableDisableEventManger()
+    public function testEnableDisableEventManger(): void
     {
         $this->assertFalse($this->translator->isEventManagerEnabled(), 'Default value');
 
@@ -355,10 +352,10 @@ class TranslatorTest extends TestCase
         $this->assertFalse($this->translator->isEventManagerEnabled());
     }
 
-    public function testEnableEventMangerViaFactory()
+    public function testEnableEventMangerViaFactory(): void
     {
         $translator = Translator::factory([
-            'event_manager_enabled' => true
+            'event_manager_enabled' => true,
         ]);
         $this->assertTrue($translator->isEventManagerEnabled());
 
@@ -366,7 +363,7 @@ class TranslatorTest extends TestCase
         $this->assertFalse($translator->isEventManagerEnabled());
     }
 
-    public function testMissingTranslationEvent()
+    public function testMissingTranslationEvent(): void
     {
         $actualEvent = null;
 
@@ -382,7 +379,7 @@ class TranslatorTest extends TestCase
 
         $this->translator->translate('foo', 'bar', 'baz');
 
-        $this->assertInstanceOf('Laminas\EventManager\Event', $actualEvent);
+        $this->assertInstanceOf(Event::class, $actualEvent);
         $this->assertEquals(
             [
                 'message'     => 'foo',
@@ -399,7 +396,7 @@ class TranslatorTest extends TestCase
         $this->assertNull($actualEvent);
     }
 
-    public function testListenerOnMissingTranslationEventCanReturnString()
+    public function testListenerOnMissingTranslationEventCanReturnString(): void
     {
         $trigger     = null;
         $doNotTriger = null;
@@ -437,7 +434,7 @@ class TranslatorTest extends TestCase
         $this->assertNull($doNotTrigger);
     }
 
-    public function testNoMessagesLoadedEvent()
+    public function testNoMessagesLoadedEvent(): void
     {
         $actualEvent = null;
 
@@ -450,7 +447,7 @@ class TranslatorTest extends TestCase
 
         $this->translator->translate('foo', 'bar', 'baz');
 
-        $this->assertInstanceOf('Laminas\EventManager\Event', $actualEvent);
+        $this->assertInstanceOf(Event::class, $actualEvent);
         $this->assertEquals(
             [
                 'locale'      => 'baz',
@@ -466,7 +463,7 @@ class TranslatorTest extends TestCase
         $this->assertNull($actualEvent);
     }
 
-    public function testListenerOnNoMessagesLoadedEventCanReturnTextDomainObject()
+    public function testListenerOnNoMessagesLoadedEventCanReturnTextDomainObject(): void
     {
         $trigger      = null;
         $doNotTrigger = null;
@@ -508,7 +505,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('BOOYAH', $result);
     }
 
-    public function testGetAllMessagesLoadedInTranslator()
+    public function testGetAllMessagesLoadedInTranslator(): void
     {
         $this->translator->setLocale('en_EN');
         $this->translator->addTranslationFile(
@@ -524,7 +521,7 @@ class TranslatorTest extends TestCase
         $this->assertEquals('Message 1 (en)', $allMessages['Message 1']);
     }
 
-    public function testGetAllMessagesReturnsNullWhenGivenTextDomainIsNotFound()
+    public function testGetAllMessagesReturnsNullWhenGivenTextDomainIsNotFound(): void
     {
         $this->translator->setLocale('en_EN');
         $this->translator->addTranslationFile(
@@ -538,7 +535,7 @@ class TranslatorTest extends TestCase
         $this->assertNull($allMessages);
     }
 
-    public function testGetAllMessagesReturnsNullWhenGivenLocaleNotExist()
+    public function testGetAllMessagesReturnsNullWhenGivenLocaleNotExist(): void
     {
         $this->translator->setLocale('en_EN');
         $this->translator->addTranslationFile(
@@ -552,17 +549,16 @@ class TranslatorTest extends TestCase
         $this->assertNull($allMessages);
     }
 
-    /**
-     * @group 33
-     */
-    public function testNullMessageArgumentShouldReturnAnEmptyString()
+    public function testNullMessageArgumentShouldReturnAnEmptyString(): void
     {
-        $loader = new TestLoader();
+        $loader             = new TestLoader();
         $loader->textDomain = new TextDomain(['foo' => 'bar']);
-        $config = new Config(['services' => [
-            'test' => $loader
-        ]]);
-        $pm = $this->translator->getPluginManager();
+        $config             = new Config([
+            'services' => [
+                'test' => $loader,
+            ],
+        ]);
+        $pm                 = $this->translator->getPluginManager();
         $config->configureServiceManager($pm);
         $this->translator->setPluginManager($pm);
         $this->translator->addTranslationFile('test', null);
