@@ -1,38 +1,27 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\View\Helper;
 
 use Laminas\I18n\View\Helper\CurrencyFormat as CurrencyFormatHelper;
+use LaminasTest\I18n\TestCase;
 use Locale;
-use PHPUnit\Framework\TestCase;
 
-/**
- * @group      Laminas_View
- * @group      Laminas_View_Helper
- */
+use function str_replace;
+
 class CurrencyFormatTest extends TestCase
 {
-    /**
-     * @var CurrencyFormatHelper
-     */
-    public $helper;
+    private CurrencyFormatHelper $helper;
 
-    /**
-     * Sets up the fixture, for example, open a network connection.
-     * This method is called before a test is executed.
-     *
-     * @return void
-     */
     protected function setUp(): void
     {
-        if (! extension_loaded('intl')) {
-            $this->markTestSkipped('ext/intl not enabled');
-        }
-
+        parent::setUp();
         $this->helper = new CurrencyFormatHelper();
     }
 
-    public function currencyProvider()
+    /** @return array<array-key, array{0: string, 1: string, 2: bool, 3: float, 4:string|null, 5: string}> */
+    public function currencyProvider(): array
     {
         return [
             //    locale   currency     show decimals       number      currencyPattern             expected
@@ -61,8 +50,14 @@ class CurrencyFormatTest extends TestCase
     /**
      * @dataProvider currencyProvider
      */
-    public function testBasic($locale, $currencyCode, $showDecimals, $number, $currencyPattern, $expected)
-    {
+    public function testBasic(
+        string $locale,
+        string $currencyCode,
+        bool $showDecimals,
+        float $number,
+        ?string $currencyPattern,
+        string $expected
+    ): void {
         $this->assertMbStringEquals(
             $expected,
             $this->helper->__invoke(
@@ -79,13 +74,13 @@ class CurrencyFormatTest extends TestCase
      * @dataProvider currencyProvider
      */
     public function testSettersProvideDefaults(
-        $locale,
-        $currencyCode,
-        $showDecimals,
-        $number,
-        $currencyPattern,
-        $expected
-    ) {
+        string $locale,
+        string $currencyCode,
+        bool $showDecimals,
+        float $number,
+        ?string $currencyPattern,
+        string $expected
+    ): void {
         $this->helper
             ->setLocale($locale)
             ->setShouldShowDecimals($showDecimals)
@@ -95,7 +90,7 @@ class CurrencyFormatTest extends TestCase
         $this->assertMbStringEquals($expected, $this->helper->__invoke($number));
     }
 
-    public function testViewhelperExecutedSequentially()
+    public function testViewHelperExecutedSequentially(): void
     {
         $helper = $this->helper;
         $helper->setShouldShowDecimals(true);
@@ -105,15 +100,15 @@ class CurrencyFormatTest extends TestCase
         $this->assertEquals('1.234,43 €', $helper(1234.4321, 'EUR', null, 'de_DE'));
     }
 
-    public function testDefaultLocale()
+    public function testDefaultLocale(): void
     {
         $this->assertMbStringEquals(Locale::getDefault(), $this->helper->getLocale());
     }
 
-    public function assertMbStringEquals($expected, $test, $message = '')
+    public function assertMbStringEquals(string $expected, string $test, string $message = ''): void
     {
         $expected = str_replace(["\xC2\xA0", ' '], '', $expected);
-        $test = str_replace(["\xC2\xA0", ' '], '', $test);
+        $test     = str_replace(["\xC2\xA0", ' '], '', $test);
         $this->assertEquals($expected, $test, $message);
     }
 }
