@@ -6,11 +6,14 @@ use Laminas\Stdlib\StringUtils;
 use Locale;
 use Traversable;
 
+use function in_array;
+use function is_array;
+use function is_scalar;
+use function preg_replace;
+
 class Alnum extends AbstractLocale
 {
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $options = [
         'locale'            => null,
         'allow_white_space' => false,
@@ -79,11 +82,13 @@ class Alnum extends AbstractLocale
             $pattern = '/[^a-zA-Z0-9' . $whiteSpace . ']/';
         } elseif (in_array($language, ['ja', 'ko', 'zh'], true)) {
             // Use english alphabet
-            $pattern = '/[^a-zA-Z0-9'  . $whiteSpace . ']/u';
+            $pattern = '/[^a-zA-Z0-9' . $whiteSpace . ']/u';
         } else {
             // Use native language alphabet
             $pattern = '/[^\p{L}\p{N}' . $whiteSpace . ']/u';
         }
+
+        $value = is_scalar($value) ? (string) $value : $value;
 
         return preg_replace($pattern, '', $value);
     }
