@@ -7,6 +7,17 @@ use Laminas\I18n\Translator\Plural\Rule as PluralRule;
 use Laminas\I18n\Translator\TextDomain;
 use Laminas\Stdlib\ErrorHandler;
 
+use function array_shift;
+use function explode;
+use function fclose;
+use function fopen;
+use function fread;
+use function fseek;
+use function sprintf;
+use function strtolower;
+use function trim;
+use function unpack;
+
 /**
  * Gettext loader.
  */
@@ -30,6 +41,7 @@ class Gettext extends AbstractFileLoader
      * load(): defined by FileLoaderInterface.
      *
      * @see    FileLoaderInterface::load()
+     *
      * @param  string $locale
      * @param  string $filename
      * @return TextDomain
@@ -49,7 +61,7 @@ class Gettext extends AbstractFileLoader
 
         ErrorHandler::start();
         $this->file = fopen($resolvedFile, 'rb');
-        $error = ErrorHandler::stop();
+        $error      = ErrorHandler::stop();
         if (false === $this->file) {
             throw new Exception\InvalidArgumentException(sprintf(
                 'Could not open file %s for reading',
@@ -73,7 +85,7 @@ class Gettext extends AbstractFileLoader
         }
 
         // Verify major revision (only 0 and 1 supported)
-        $majorRevision = ($this->readInteger() >> 16);
+        $majorRevision = $this->readInteger() >> 16;
 
         if ($majorRevision !== 0 && $majorRevision !== 1) {
             fclose($this->file);
@@ -136,7 +148,7 @@ class Gettext extends AbstractFileLoader
             $rawHeaders = explode("\n", trim($textDomain['']));
 
             foreach ($rawHeaders as $rawHeader) {
-                list($header, $content) = explode(':', $rawHeader, 2);
+                [$header, $content] = explode(':', $rawHeader, 2);
 
                 if (strtolower(trim($header)) === 'plural-forms') {
                     $textDomain->setPluralRule(PluralRule::fromString($content));
