@@ -7,6 +7,15 @@ use Laminas\I18n\Exception;
 use Laminas\I18n\Translator\Plural\Rule as PluralRule;
 use Laminas\I18n\Translator\TextDomain;
 
+use function array_shift;
+use function count;
+use function gettype;
+use function is_array;
+use function is_file;
+use function is_readable;
+use function sprintf;
+use function stream_resolve_include_path;
+
 /**
  * PHP INI format loader.
  */
@@ -16,6 +25,7 @@ class Ini extends AbstractFileLoader
      * load(): defined by FileLoaderInterface.
      *
      * @see    FileLoaderInterface::load()
+     *
      * @param  string $locale
      * @param  string $filename
      * @return TextDomain
@@ -24,7 +34,7 @@ class Ini extends AbstractFileLoader
     public function load($locale, $filename)
     {
         $resolvedIncludePath = stream_resolve_include_path($filename);
-        $fromIncludePath = ($resolvedIncludePath !== false) ? $resolvedIncludePath : $filename;
+        $fromIncludePath     = $resolvedIncludePath !== false ? $resolvedIncludePath : $filename;
         if (! $fromIncludePath || ! is_file($fromIncludePath) || ! is_readable($fromIncludePath)) {
             throw new Exception\InvalidArgumentException(sprintf(
                 'Could not find or open file %s for reading',
