@@ -6,6 +6,13 @@ use Laminas\I18n\Exception;
 use Laminas\I18n\Translator\Plural\Rule as PluralRule;
 use Laminas\I18n\Translator\TextDomain;
 
+use function gettype;
+use function is_array;
+use function is_file;
+use function is_readable;
+use function sprintf;
+use function stream_resolve_include_path;
+
 /**
  * PHP array loader.
  */
@@ -15,6 +22,7 @@ class PhpArray extends AbstractFileLoader
      * load(): defined by FileLoaderInterface.
      *
      * @see    FileLoaderInterface::load()
+     *
      * @param  string $locale
      * @param  string $filename
      * @return TextDomain
@@ -23,7 +31,7 @@ class PhpArray extends AbstractFileLoader
     public function load($locale, $filename)
     {
         $resolvedIncludePath = stream_resolve_include_path($filename);
-        $fromIncludePath = ($resolvedIncludePath !== false) ? $resolvedIncludePath : $filename;
+        $fromIncludePath     = $resolvedIncludePath !== false ? $resolvedIncludePath : $filename;
         if (! $fromIncludePath || ! is_file($fromIncludePath) || ! is_readable($fromIncludePath)) {
             throw new Exception\InvalidArgumentException(sprintf(
                 'Could not find or open file %s for reading',
