@@ -4,11 +4,13 @@ namespace Laminas\I18n\Validator;
 
 use Laminas\I18n\Filter\Alpha as AlphaFilter;
 
+use function is_string;
+
 class Alpha extends Alnum
 {
-    const INVALID      = 'alphaInvalid';
-    const NOT_ALPHA    = 'notAlpha';
-    const STRING_EMPTY = 'alphaStringEmpty';
+    public const INVALID      = 'alphaInvalid';
+    public const NOT_ALPHA    = 'notAlpha';
+    public const STRING_EMPTY = 'alphaStringEmpty';
 
     /**
      * Alphabetic filter used for validation
@@ -25,7 +27,7 @@ class Alpha extends Alnum
     protected $messageTemplates = [
         self::INVALID      => 'Invalid type given. String expected',
         self::NOT_ALPHA    => 'The input contains non alphabetic characters',
-        self::STRING_EMPTY => 'The input is an empty string'
+        self::STRING_EMPTY => 'The input is an empty string',
     ];
 
     /**
@@ -34,7 +36,7 @@ class Alpha extends Alnum
      * @var array
      */
     protected $options = [
-        'allowWhiteSpace' => false,  // Whether to allow white space characters; off by default
+        'allowWhiteSpace' => false, // Whether to allow white space characters; off by default
     ];
 
     /**
