@@ -5,9 +5,11 @@ namespace Laminas\I18n\View\Helper;
 use DateTimeInterface;
 use IntlCalendar;
 use IntlDateFormatter;
-use Laminas\I18n\Exception;
 use Laminas\View\Helper\AbstractHelper;
 use Locale;
+
+use function date_default_timezone_get;
+use function md5;
 
 /**
  * View helper for formatting dates.
@@ -31,22 +33,9 @@ class DateFormat extends AbstractHelper
     /**
      * Formatter instances
      *
-     * @var array
+     * @var array<string, IntlDateFormatter>
      */
     protected $formatters = [];
-
-    /**
-     * @throws Exception\ExtensionNotLoadedException if ext/intl is not present
-     */
-    public function __construct()
-    {
-        if (! extension_loaded('intl')) {
-            throw new Exception\ExtensionNotLoadedException(sprintf(
-                '%s component requires the intl PHP extension',
-                __NAMESPACE__
-            ));
-        }
-    }
 
     /**
      * Format a date
@@ -70,7 +59,7 @@ class DateFormat extends AbstractHelper
         }
 
         $timezone    = $this->getTimezone();
-        $formatterId = md5($dateType . "\0" . $timeType . "\0" . $locale . "\0" . $pattern . "\0" . $timezone);
+        $formatterId = md5($dateType . "\0" . $timeType . "\0" . $locale . "\0" . (string) $pattern . "\0" . $timezone);
 
         if (! isset($this->formatters[$formatterId])) {
             $this->formatters[$formatterId] = new IntlDateFormatter(
@@ -79,7 +68,7 @@ class DateFormat extends AbstractHelper
                 $timeType,
                 $timezone,
                 IntlDateFormatter::GREGORIAN,
-                $pattern
+                $pattern ?? ''
             );
         }
 
@@ -120,13 +109,9 @@ class DateFormat extends AbstractHelper
      */
     public function setTimezone($timezone)
     {
-        $this->timezone = (string) $timezone;
-
-        // The method setTimeZoneId is deprecated as of PHP 5.5.0
-        $setTimeZoneMethodName = (PHP_VERSION_ID < 50500) ? 'setTimeZoneId' : 'setTimeZone';
-
+        $this->timezone = $timezone;
         foreach ($this->formatters as $formatter) {
-            $formatter->$setTimeZoneMethodName($this->timezone);
+            $formatter->setTimeZone($this->timezone);
         }
 
         return $this;
