@@ -2,10 +2,14 @@
 
 namespace Laminas\I18n\View\Helper;
 
-use Laminas\I18n\Exception;
 use Laminas\View\Helper\AbstractHelper;
 use Locale;
 use NumberFormatter;
+
+use function is_array;
+use function is_int;
+use function md5;
+use function serialize;
 
 /**
  * View helper for formatting dates.
@@ -55,19 +59,6 @@ class NumberFormat extends AbstractHelper
     protected $locale;
 
     /**
-     * @throws Exception\ExtensionNotLoadedException if ext/intl is not present
-     */
-    public function __construct()
-    {
-        if (! extension_loaded('intl')) {
-            throw new Exception\ExtensionNotLoadedException(sprintf(
-                '%s component requires the intl PHP extension',
-                __NAMESPACE__
-            ));
-        }
-    }
-
-    /**
      * Format a number
      *
      * @param  int|float   $number
@@ -84,7 +75,7 @@ class NumberFormat extends AbstractHelper
         $formatType = null,
         $locale = null,
         $decimals = null,
-        array $textAttributes = null
+        ?array $textAttributes = null
     ) {
         if (null === $locale) {
             $locale = $this->getLocale();
