@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\Translator;
 
 use Laminas\Filter\FilterInterface;
@@ -7,28 +9,28 @@ use Laminas\I18n\Exception\RuntimeException;
 use Laminas\I18n\Translator\LoaderPluginManager;
 use Laminas\ServiceManager\ServiceManager;
 use Laminas\ServiceManager\Test\CommonPluginManagerTrait;
-use PHPUnit\Framework\TestCase;
+use LaminasTest\I18n\TestCase;
 
 class LoaderPluginManagerCompatibilityTest extends TestCase
 {
     use CommonPluginManagerTrait;
 
-    protected function getPluginManager()
+    protected function getPluginManager(): LoaderPluginManager
     {
         return new LoaderPluginManager(new ServiceManager());
     }
 
-    protected function getV2InvalidPluginException()
+    protected function getV2InvalidPluginException(): string
     {
         return RuntimeException::class;
     }
 
-    protected function getInstanceOf()
+    protected function getInstanceOf(): string
     {
         return FilterInterface::class;
     }
 
-    public function testInstanceOfMatches()
+    public function testInstanceOfMatches(): void
     {
         $this->markTestSkipped('Test skipped as LoaderPluginManager allows multiple instance types');
     }
