@@ -1,13 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
 namespace LaminasTest\I18n\Translator\Plural;
 
 use Laminas\I18n\Translator\Plural\Rule;
-use PHPUnit\Framework\TestCase;
+use LaminasTest\I18n\TestCase;
 
 class RuleTest extends TestCase
 {
-    public static function parseRuleProvider()
+    /** @return array<string, array{0: string, 1: int}> */
+    public function parseRuleProvider(): array
     {
         return [
             // Basic calculations
@@ -46,14 +49,14 @@ class RuleTest extends TestCase
             'boolean-or-false'  => ['0 || 0', 0],
 
             // Variable injection
-            'variable-injection' => ['n', 0]
+            'variable-injection' => ['n', 0],
         ];
     }
 
     /**
      * @dataProvider parseRuleProvider
      */
-    public function testParseRules($rule, $expectedValue)
+    public function testParseRules(string $rule, int $expectedValue): void
     {
         $this->assertEquals(
             $expectedValue,
@@ -61,7 +64,8 @@ class RuleTest extends TestCase
         );
     }
 
-    public static function completeRuleProvider()
+    /** @return array<array-key, array{0: string, 1: string}> */
+    public function completeRuleProvider(): array
     {
         // Taken from original gettext tests
         return [
@@ -70,63 +74,63 @@ class RuleTest extends TestCase
                 '10111111111111111111111111111111111111111111111111111111111111'
                 . '111111111111111111111111111111111111111111111111111111111111'
                 . '111111111111111111111111111111111111111111111111111111111111'
-                . '111111111111111111'
+                . '111111111111111111',
             ],
             [
                 'n>1',
                 '00111111111111111111111111111111111111111111111111111111111111'
                 . '111111111111111111111111111111111111111111111111111111111111'
                 . '111111111111111111111111111111111111111111111111111111111111'
-                . '111111111111111111'
+                . '111111111111111111',
             ],
             [
                 'n==1 ? 0 : n==2 ? 1 : 2',
                 '20122222222222222222222222222222222222222222222222222222222222'
                 . '222222222222222222222222222222222222222222222222222222222222'
                 . '222222222222222222222222222222222222222222222222222222222222'
-                . '222222222222222222'
+                . '222222222222222222',
             ],
             [
                 'n==1 ? 0 : (n==0 || (n%100 > 0 && n%100 < 20)) ? 1 : 2',
                 '10111111111111111111222222222222222222222222222222222222222222'
                 . '222222222222222222222222222222222222222111111111111111111122'
                 . '222222222222222222222222222222222222222222222222222222222222'
-                . '222222222222222222'
+                . '222222222222222222',
             ],
             [
                 'n%10==1 && n%100!=11 ? 0 : n%10>=2 && (n%100<10 || n%100>=20) ? 1 : 2',
                 '20111111112222222222201111111120111111112011111111201111111120'
                 . '111111112011111111201111111120111111112011111111222222222220'
                 . '111111112011111111201111111120111111112011111111201111111120'
-                . '111111112011111111'
+                . '111111112011111111',
             ],
             [
                 'n%10==1 && n%100!=11 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2',
                 '20111222222222222222201112222220111222222011122222201112222220'
                 . '111222222011122222201112222220111222222011122222222222222220'
                 . '111222222011122222201112222220111222222011122222201112222220'
-                . '111222222011122222'
+                . '111222222011122222',
             ],
             [
                 'n%100/10==1 ? 2 : n%10==1 ? 0 : (n+9)%10>3 ? 2 : 1',
                 '20111222222222222222201112222220111222222011122222201112222220'
                 . '111222222011122222201112222220111222222011122222222222222220'
                 . '111222222011122222201112222220111222222011122222201112222220'
-                . '111222222011122222'
+                . '111222222011122222',
             ],
             [
                 'n==1 ? 0 : n%10>=2 && n%10<=4 && (n%100<10 || n%100>=20) ? 1 : 2',
                 '20111222222222222222221112222222111222222211122222221112222222'
                 . '111222222211122222221112222222111222222211122222222222222222'
                 . '111222222211122222221112222222111222222211122222221112222222'
-                . '111222222211122222'
+                . '111222222211122222',
             ],
             [
                 'n%100==1 ? 0 : n%100==2 ? 1 : n%100==3 || n%100==4 ? 2 : 3',
                 '30122333333333333333333333333333333333333333333333333333333333'
                 . '333333333333333333333333333333333333333012233333333333333333'
                 . '333333333333333333333333333333333333333333333333333333333333'
-                . '333333333333333333'
+                . '333333333333333333',
             ],
         ];
     }
@@ -134,7 +138,7 @@ class RuleTest extends TestCase
     /**
      * @dataProvider completeRuleProvider
      */
-    public function testCompleteRules($rule, $expectedValues)
+    public function testCompleteRules(string $rule, string $expectedValues): void
     {
         $rule = Rule::fromString('nplurals=9; plural=' . $rule);
 
@@ -143,7 +147,7 @@ class RuleTest extends TestCase
         }
     }
 
-    public function testGetNumPlurals()
+    public function testGetNumPlurals(): void
     {
         $rule = Rule::fromString('nplurals=9; plural=n');
         $this->assertEquals(9, $rule->getNumPlurals());
